+++ test/integration/patch-resource/success.js
@@ -176,79 +176,6 @@ describe('Resource PATCH success', function() {
     });
   });
 
-  describe('when the request is valid, with a relationship', () => {
-    beforeEach((done) => {
-      this.options = {
-        resourcesDirectory: path.join(fixturesDirectory, 'kitchen-sink'),
-        apiVersion: 24
-      };
-
-      const relationGuestSeeds = [
-        {first_name: 'sandwiches'},
-        {first_name: 'what'},
-        {first_name: 'pls'}
-      ];
-
-      const relationSeeds = [
-        {name: 'james', owner_id: '1'}
-      ];
-
-      applyMigrations(this.options)
-        .then(() => seed('relation_guest', relationGuestSeeds))
-        .then(() => seed('relation', relationSeeds))
-        .then(() => done());
-    });
-
-    it('should return a 200 OK, with the updated resource', (done) => {
-      const expectedData = {
-        type: 'relations',
-        id: '1',
-        attributes: {
-          name: 'james',
-          size: null
-        },
-        relationships: {
-          owner: {
-            data: {
-              id: '3',
-              type: 'relation_guests'
-            },
-            links: {
-              self: '/v24/relations/1/relationships/owner',
-              related: '/v24/relations/1/owner'
-            }
-          }
-        }
-      };
-
-      const expectedLinks = {
-        self: '/v24/relations/1'
-      };
-
-      request(app(this.options))
-        .patch('/v24/relations/1')
-        .send({
-          data: {
-            id: '1',
-            type: 'relations',
-            relationships: {
-              owner: {
-                data: {
-                  id: '3',
-                  type: 'relation_guests'
-                }
-              }
-            }
-          }
-        })
-        .expect(validators.basicValidation)
-        .expect(validators.assertData(expectedData))
-        .expect(validators.assertLinks(expectedLinks))
-        .expect(200)
-        .end(done);
-    });
-  });
-
   describe('when the request leaves out non-nullable data in a patch', () => {
     beforeEach((done) => {
       this.options = {
